+++ test/functional-tests/node-transformers/dead-code-injection-transformers/fixtures/function-declaration-inside-block-statement.js
@@ -1,0 +1,25 @@
+(function(){
+    if (true) {
+        var foo = function () {
+            console.log('abc');
+        };
+        var bar = function () {
+            console.log('def');
+        };
+        var baz = function () {
+            console.log('ghi');
+        };
+        var bark = function () {
+            console.log('jkl');
+        };
+
+        if (true) {
+            function hawk (param) {}
+        }
+
+        foo();
+        bar();
+        baz();
+        bark();
+    }
+})();