+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v1.3.0
 ---
 * Improvements of `stringArrayEncoding`: `base64` and `rc4`
 * **CLI**: added config file extension validation (it still supports `.js` and `.json` extensions)
+* Fixed https://github.com/javascript-obfuscator/javascript-obfuscator/issues/499
 
 v1.2.2
 ---